move specialized regression tests
