+++ astropy/cosmology/tests/test_cosmology.py
@@ -1242,67 +1242,3 @@ def test_massivenu_density():
     ztest = ztest.astype(int)
     assert allclose(tcos.nu_relative_density(ztest), nurel_exp, rtol=5e-3)
     assert allclose(tcos.Onu(ztest), onu_exp, rtol=5e-3)
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
-def test_elliptic_comoving_distance_z1z2():
-    """Regression test for #8388."""
-    cosmo = flrw.LambdaCDM(70.0, 2.3, 0.05, Tcmb0=0)
-    z = 0.2
-    assert allclose(
-        cosmo.comoving_distance(z), cosmo._integral_comoving_distance_z1z2(0.0, z)
-    )
-    assert allclose(
-        cosmo._elliptic_comoving_distance_z1z2(0.0, z),
-        cosmo._integral_comoving_distance_z1z2(0.0, z),
-    )
-
-
-SPECIALIZED_COMOVING_DISTANCE_COSMOLOGIES = [
-    flrw.FlatLambdaCDM(H0=70, Om0=0.0, Tcmb0=0.0),  # de Sitter
-    flrw.FlatLambdaCDM(H0=70, Om0=1.0, Tcmb0=0.0),  # Einstein - de Sitter
-    flrw.FlatLambdaCDM(H0=70, Om0=0.3, Tcmb0=0.0),  # Hypergeometric
-    flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.6, Tcmb0=0.0),  # Elliptic
-]
-
-
-ITERABLE_REDSHIFTS = [
-    (0, 1, 2, 3, 4),  # tuple
-    [0, 1, 2, 3, 4],  # list
-    np.array([0, 1, 2, 3, 4]),  # array
-]
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
-@pytest.mark.parametrize("cosmo", SPECIALIZED_COMOVING_DISTANCE_COSMOLOGIES)
-@pytest.mark.parametrize("z", ITERABLE_REDSHIFTS)
-def test_comoving_distance_iterable_argument(cosmo, z):
-    """
-    Regression test for #10980
-    Test that specialized comoving distance methods handle iterable arguments.
-    """
-
-    assert allclose(
-        cosmo.comoving_distance(z), cosmo._integral_comoving_distance_z1z2(0.0, z)
-    )
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
-@pytest.mark.parametrize("cosmo", SPECIALIZED_COMOVING_DISTANCE_COSMOLOGIES)
-def test_comoving_distance_broadcast(cosmo):
-    """
-    Regression test for #10980
-    Test that specialized comoving distance methods broadcast array arguments.
-    """
-
-    z1 = np.zeros((2, 5))
-    z2 = np.ones((3, 1, 5))
-    z3 = np.ones((7, 5))
-    output_shape = np.broadcast(z1, z2).shape
-
-    # Check compatible array arguments return an array with the correct shape
-    assert cosmo._comoving_distance_z1z2(z1, z2).shape == output_shape
-
-    # Check incompatible array arguments raise an error
-    with pytest.raises(ValueError, match="z1 and z2 have different shapes"):
-        cosmo._comoving_distance_z1z2(z1, z3)
